+++ fam/llm/finetune.py
@@ -26,9 +26,9 @@ from fam.telemetry.posthog import PosthogClient
 # see fam/telemetry/README.md for more information
 posthog = PosthogClient()
 
-dtype: Literal["bfloat16", "float16", "tfloat32", "float32"] = (
-    "bfloat16" if torch.cuda.is_available() and torch.cuda.is_bf16_supported() else "float16"
-)  # 'float32', 'bfloat16', or 'float16', the latter will auto implement a GradScaler
+dtype: Literal["float16", "float16", "tfloat32", "float32"] = (
+    "float16" if torch.cuda.is_available() and torch.cuda.is_bf16_supported() else "float16"
+)  # 'float32', 'float16', or 'float16', the latter will auto implement a GradScaler
 seed_offset = 0
 
 torch.manual_seed(seed + seed_offset)
@@ -36,7 +36,7 @@ torch.backends.cuda.matmul.allow_tf32 = True if dtype != "float32" else False
 torch.backends.cudnn.allow_tf32 = True if dtype != "float32" else False
 device_type = "cuda" if "cuda" in device else "cpu"  # for later use in torch.autocast
 # note: float16 data type will automatically use a GradScaler
-ptdtype = {"float32": torch.float32, "tfloat32": torch.float32, "bfloat16": torch.bfloat16, "float16": torch.float16}[
+ptdtype = {"float32": torch.float32, "tfloat32": torch.float32, "float16": torch.float16, "float16": torch.float16}[
     dtype
 ]
 ctx = nullcontext() if device_type == "cpu" else torch.amp.autocast(device_type=device_type, dtype=ptdtype)
